+++ app/controllers/aip/v1/entities_controller.rb
@@ -4,9 +4,10 @@ class Aip::V1::EntitiesController < ApplicationController
   THESIS_CLASS_NAME = Thesis.name.freeze
   ITEM_INSTITUTIONAL_REPOSITORY_NAME = 'IRItem'.freeze
   THESIS_INSTITUTIONAL_REPOSITORY_NAME = 'IRThesis'.freeze
+  FILESET_INSTITUTIONAL_REPOSITORY_NAME = 'IRFileSet'.freeze
   ENTITY_INSTITUTIONAL_REPOSITORY_NAME = 'IREntity'.freeze
 
-  before_action :load_entity, only: [:show, :file_sets, :file_paths]
+  before_action :load_entity, only: [:show, :file_sets, :file_paths, :file_set]
   before_action :load_file, only: [:file_set, :fixity_file, :original_file]
   before_action :ensure_access
 
@@ -41,7 +42,7 @@ class Aip::V1::EntitiesController < ApplicationController
     # but are not directly referenced by a predicate in the system because they
     # require some level of processing to obtain.
 
-    rdf_graph_creator.insert(
+    rdf_graph_creator.graph.insert(
       owner_email_statement,
       entity_model_statement,
       *entity_type_statements,
@@ -64,14 +65,25 @@ class Aip::V1::EntitiesController < ApplicationController
       RDF::Vocab::BIBO.contributorList
     )
 
-    # Handle special case where predicate http://projecthydra.org/ns/auth/acl#embargoHistory
-    # needs to maintain the order in which it was entered. The original nodes
-    # the predicate are removed and a new rdf list is inserted instead with the
-    # predicate
+    # Thesis predicates that need to be added as sorted lists
+    rdf_graph_creator.copy_predicate_to_sorted_rdf_list(
+      self_subject,
+      TERMS[:ual].department,
+      TERMS[:ual].department_list
+    )
+    rdf_graph_creator.copy_predicate_to_sorted_rdf_list(
+      self_subject,
+      TERMS[:ual].supervisor,
+      TERMS[:ual].supervisor_list
+    )
+
+    # Handle special case where predicate http://projecthydra.org/ns/auth/acl#embargoHistory needs to maintain the order
+    # in which it was entered. The original nodes the predicate are removed and a new rdf list is inserted instead with
+    # the predicate
 
     rdf_graph_creator.replace_predicate_with_sorted_rdf_list(self_subject, ::TERMS[:acl].embargo_history)
 
-    render plain: rdf_graph_creator.to_n3, status: :ok
+    render plain: rdf_graph_creator.graph.to_n3, status: :ok
   end
 
   def file_sets
@@ -112,11 +124,50 @@ class Aip::V1::EntitiesController < ApplicationController
     ActsAsRdfable.add_annotation_bindings!(@file.blob)
 
     rdf_graph_creator = RdfGraphCreationService.new(@file.blob, prefixes, self_subject)
-    rdf_graph_creator.insert(
-      RDF::Statement(subject: self_subject, predicate: RDF.type, object: RDF::Vocab::PCDM.Object)
+
+    file_view_uri = RDF::URI(
+      file_view_item_url(
+        id: @entity.id,
+        file_set_id: @file.fileset_uuid,
+        file_name: @file.filename.to_s
+      )
     )
 
-    render plain: rdf_graph_creator.to_n3, status: :ok
+    statements = [
+      owner_email_statement,
+      *entity_member_of_statements,
+      RDF::Statement(subject: self_subject, predicate: RDF.type, object: RDF::Vocab::PCDM.Object),
+      RDF::Statement(subject: self_subject, predicate: RDF.type, object: RDF::Vocab::Fcrepo4.Resource),
+      RDF::Statement(subject: self_subject, predicate: RDF.type, object: RDF::Vocab::Fcrepo4.Container),
+      RDF::Statement(subject: self_subject, predicate: RDF::Vocab::DC.accessRights, object: @entity.visibility),
+      RDF::Statement(subject: self_subject, predicate: RDF::Vocab::EBUCore.dateIngested, object: @entity.date_ingested),
+      RDF::Statement(subject: self_subject, predicate: RDF::Vocab::Fcrepo4.hasParent, object: self_subject.parent),
+      RDF::Statement(
+        subject: self_subject,
+        predicate: RDF::Vocab::PCDM.hasFile,
+        object: self_subject / 'original_file'
+      ),
+      RDF::Statement(
+        subject: self_subject,
+        predicate: TERMS[:ual].record_created_in_jupiter,
+        object: @entity.record_created_at
+      ),
+      RDF::Statement(
+        subject: self_subject,
+        predicate: ::TERMS[:fedora].has_model,
+        object: FILESET_INSTITUTIONAL_REPOSITORY_NAME
+      ),
+      RDF::Statement(
+        subject: self_subject,
+        predicate: ::TERMS[:ual].sitemap_link,
+        object: RDF::URI('rs:ln') + %Q( href="#{file_view_uri.path}" rel="content" hash="md5:#{@file.checksum}" ) +
+                %Q(length="#{@file.byte_size}" type="#{@file.content_type}")
+      )
+    ]
+
+    rdf_graph_creator.graph.insert(*statements)
+
+    render plain: rdf_graph_creator.graph.to_n3, status: :ok
   end
 
   def fixity_file
@@ -131,19 +182,17 @@ class Aip::V1::EntitiesController < ApplicationController
       @file.blob, prefixes, subject
     )
 
-    statement_definitions = [
-      { subject: subject, predicate: RDF::Vocab::PREMIS.hasFixity, object: self_subject },
-      { subject: self_subject, predicate: RDF::Vocab::PREMIS.hasEventOutcome, object: 'SUCCESS' },
-      { subject: self_subject, predicate: RDF::Vocab::PREMIS.hasMessageDigestAlgorithm, object: 'md5' },
-      { subject: self_subject, predicate: RDF.type, object: RDF::Vocab::PREMIS.EventOutcomeDetail },
-      { subject: self_subject, predicate: RDF.type, object: RDF::Vocab::PREMIS.Fixity }
+    statements = [
+      RDF::Statement(subject: subject, predicate: RDF::Vocab::PREMIS.hasFixity, object: self_subject),
+      RDF::Statement(subject: self_subject, predicate: RDF::Vocab::PREMIS.hasEventOutcome, object: 'SUCCESS'),
+      RDF::Statement(subject: self_subject, predicate: RDF::Vocab::PREMIS.hasMessageDigestAlgorithm, object: 'md5'),
+      RDF::Statement(subject: self_subject, predicate: RDF.type, object: RDF::Vocab::PREMIS.EventOutcomeDetail),
+      RDF::Statement(subject: self_subject, predicate: RDF.type, object: RDF::Vocab::PREMIS.Fixity)
     ]
 
-    statement_definitions.each do |statement_definition|
-      rdf_graph_creator.insert(RDF::Statement(statement_definition))
-    end
+    rdf_graph_creator.graph.insert(*statements)
 
-    render plain: rdf_graph_creator.to_n3, status: :ok
+    render plain: rdf_graph_creator.graph.to_n3, status: :ok
   end
 
   def original_file
@@ -162,9 +211,33 @@ class Aip::V1::EntitiesController < ApplicationController
     ActsAsRdfable.add_annotation_bindings!(@file.blob)
 
     rdf_graph_creator = RdfGraphCreationService.new(@file.blob, prefixes, self_subject)
-    rdf_graph_creator.insert(RDF::Statement(subject: self_subject, predicate: RDF.type, object: RDF::Vocab::PCDM.File))
+    insert_nodes = [
+      RDF::Statement(subject: self_subject, predicate: RDF.type, object: RDF::Vocab::PCDM.File),
+      RDF::Statement(
+        subject: self_subject,
+        predicate: RDF::Vocab::Fcrepo4.hasFixityService,
+        object: self_subject.parent / 'fixity'
+      ),
+      # We need to change the value for the predicate RDF::Vocab::PREMIS.hasMessageDigest so that it includes the
+      # uniform resource name with the algorightm used to create the checksum. In this case, we know that Active storage
+      # uses MD5
+      RDF::Statement(
+        subject: self_subject,
+        predicate: RDF::Vocab::PREMIS.hasMessageDigest,
+        object: RDF::URI.new('urn:md5') / @file.blob.checksum
+      ),
+      RDF::Statement(
+        subject: self_subject,
+        predicate: RDF::Vocab::Fcrepo4.hasParent,
+        object: self_subject.parent
+      )
+    ]
 
-    render plain: rdf_graph_creator.to_n3, status: :ok
+    rdf_graph_creator.graph.delete_insert(
+      rdf_graph_creator.graph.query(predicate: RDF::Vocab::PREMIS.hasMessageDigest), insert_nodes
+    )
+
+    render plain: rdf_graph_creator.graph.to_n3, status: :ok
   end
 
   def file_paths